chore: moved Contacts tests to their own folder. #19
