+++ components/Contacts/__tests__/CreateContact-test.js
@@ -1,6 +1,6 @@
 import 'react-native';
 import React from 'react';
-import CreateContact from '../Contacts/CreateContact';
+import CreateContact from '../CreateContact';
 import renderer from 'react-test-renderer';
 import ShallowRenderer from 'react-test-renderer/shallow';
